move dax modules into dedicated dir
