+++ bin/ds9_imgfit.sh
@@ -114,7 +114,7 @@ mdl1.xpos=$xx
 mdl1.ypos=$yy
 
 
-from ciao_contrib._tools.dax_model_editor import *
+from dax.dax_model_editor import *
 DaxModelEditor([mdl1,bkg1], hide_plot_button=True).run()
