+++ bin/ds9_radial_fit
@@ -24,7 +24,7 @@
 import sys
 import os
 
-from ciao_contrib._tools.dax_plot_utils import *
+from dax.dax_plot_utils import *
 
 
 def blt_plot_fit(sherpa_plot_fit_obj, xpa="ds9", 
@@ -154,7 +154,7 @@ def fit_profile( xpa, infile, model, xcol, method, ycol="sur_bri", ycol_err="sur
     sherpa.set_model(model)
     sherpa.set_method(method)
 
-    from ciao_contrib._tools.dax_model_editor import DaxModelEditor
+    from dax.dax_model_editor import DaxModelEditor
 
     s = sherpa.get_source()
     if hasattr(s,"parts") is True:
