+++ setup.py
@@ -16,6 +16,7 @@ mods = [ "ciao_contrib",
     "ciao_contrib/cda",
     "coords",
     "crates_contrib",
+    "dax",
     "sherpa_contrib",
     "sherpa_contrib/profiles",
     "sherpa_contrib/tests",
@@ -25,7 +26,7 @@ mods = [ "ciao_contrib",
 
 from distutils.core import setup
 setup( name='ciao-contrib',
-        version='4.12.0',
+        version='4.12.2',
         license='GNU GPL v3',
         description='CIAO Contributed scripts',
         author='CXCSDS and Friends',
